refactor: group/[id]
