+++ src/app/groups/components/Sidebar.tsx
