+++ src/components/features/analytics/ActivityStats/index.tsx
@@ -1,4 +1,4 @@
-import Chart from "@/components/features/analytics/Chart";
+import Chart from '@/app/groups/components/Chart';
 
 export default function ActivityStats() {
   return (
